youtubeapi/requirements.txt
APScheduler==3.9.1.post1
asgiref==3.5.2
cachetools==5.2.0
certifi==2022.12.7
charset-normalizer==2.1.1
Django==3.2
django-cors-headers==3.13.0
django-cron==0.6.0
django-crontab==0.7.1
django-filter==22.1
djangorestframework==3.14.0
google-api-core==2.11.0
google-api-python-client==2.70.0
google-auth==2.15.0
google-auth-httplib2==0.1.0
googleapis-common-protos==1.57.0
httplib2==0.21.0
idna==3.4
protobuf==4.21.12
psycopg2-binary==2.9.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
pytz==2022.6
pytz-deprecation-shim==0.1.0.post0
requests==2.28.1
rsa==4.9
six==1.16.0
sqlparse==0.4.3
tzdata==2022.7
tzlocal==4.2
uritemplate==4.1.1
urllib3==1.26.13
